server/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
geopandas==1.0.1
shapely==2.0.6
fiona==1.10.1
pyproj==3.7.0
